style(plan): add tailwind class
